backend/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
SQLAlchemy==2.0.4
Werkzeug==2.2.3
celery==5.2.7
redis==4.5.1
python-dotenv==1.0.0
PyJWT==2.6.0
Pillow==9.4.0
email-validator==2.0.0
gunicorn==20.1.0
Jinja2==3.1.2
MarkupSafe==2.1.2
itsdangerous==2.1.2
click==8.1.3
blinker==1.5
alembic==1.9.4 